+++ app/_components/ui/templates/form.tsx
@@ -10,7 +10,7 @@ import {
   useFormContext,
 } from 'react-hook-form';
 
-import { cn } from '@/app/utils/utils';
+import { cn } from '@/lib/utils';
 import { Label } from '@/components/ui/label';
 
 const Form = FormProvider;
