+++ app/_components/ui/templates/input.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { cn } from '@/app/utils/utils';
+import { cn } from '@/lib/utils';
 
 export interface InputProps
   extends React.InputHTMLAttributes<HTMLInputElement> {}
